+++ src/messaging/mod.rs
@@ -1,3 +1,5 @@
 pub mod handler;
+pub mod messages;
 
 pub use handler::*;
+pub use messages::*;